Moved Kafka wrapper tests to integration tests
