+++ distribution/server/src/assemble/LICENSE.bin.txt
@@ -387,7 +387,7 @@ The Apache Software License, Version 2.0
     - org.apache.distributedlog-distributedlog-core-4.7.2-tests.jar
     - org.apache.distributedlog-distributedlog-core-4.7.2.jar
     - org.apache.distributedlog-distributedlog-protocol-4.7.2.jar
- * LZ4 -- net.jpountz.lz4-lz4-1.3.0.jar
+ * LZ4 -- org.lz4-lz4-java-1.5.0.jar
  * AsyncHttpClient
     - org.asynchttpclient-async-http-client-2.1.0-alpha26.jar
     - org.asynchttpclient-async-http-client-netty-utils-2.1.0-alpha26.jar
