--- pulsar-client-kafka-compat/pulsar-client-kafka/src/test/java/org/apache/pulsar/client/kafka/compat/tests/KafkaProducerTest.java
@@ -1,119 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.pulsar.client.kafka.compat.tests;
-
-import static org.testng.Assert.assertEquals;
-import static org.testng.Assert.assertNull;
-
-import java.util.Properties;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
-
-import org.apache.kafka.clients.producer.Producer;
-import org.apache.kafka.clients.producer.ProducerRecord;
-import org.apache.kafka.clients.producer.PulsarKafkaProducer;
-import org.apache.kafka.common.serialization.IntegerSerializer;
-import org.apache.kafka.common.serialization.StringSerializer;
-import org.apache.pulsar.broker.service.BrokerTestBase;
-import org.apache.pulsar.client.api.Consumer;
-import org.apache.pulsar.client.api.Message;
-import org.testng.annotations.AfterClass;
-import org.testng.annotations.BeforeClass;
-import org.testng.annotations.Test;
-
-public class KafkaProducerTest extends BrokerTestBase {
-    @BeforeClass
-    @Override
-    protected void setup() throws Exception {
-        isTcpLookup = true;
-        super.baseSetup();
-    }
-
-    @AfterClass
-    @Override
-    protected void cleanup() throws Exception {
-        super.internalCleanup();
-    }
-
-    @Test
-    public void testSimpleProducer() throws Exception {
-        String topic = "persistent://prop/ns-abc/testSimpleProducer";
-
-        Consumer<byte[]> pulsarConsumer = pulsarClient.newConsumer().topic(topic).subscriptionName("my-subscription")
-                .subscribe();
-
-        Properties props = new Properties();
-        props.put("bootstrap.servers", lookupUrl.toString());
-
-        props.put("key.serializer", IntegerSerializer.class.getName());
-        props.put("value.serializer", StringSerializer.class.getName());
-
-        Producer<Integer, String> producer = new PulsarKafkaProducer<>(props);
-
-        for (int i = 0; i < 10; i++) {
-            producer.send(new ProducerRecord<Integer, String>(topic, i, "hello-" + i));
-        }
-
-        producer.flush();
-        producer.close();
-
-        for (int i = 0; i < 10; i++) {
-            Message<byte[]> msg = pulsarConsumer.receive(1, TimeUnit.SECONDS);
-            assertEquals(new String(msg.getData()), "hello-" + i);
-            pulsarConsumer.acknowledge(msg);
-        }
-    }
-
-    @Test(timeOut = 10000)
-    public void testProducerCallback() throws Exception {
-        String topic = "persistent://prop/ns-abc/testProducerCallback";
-
-        Consumer<byte[]> pulsarConsumer = pulsarClient.newConsumer().topic(topic).subscriptionName("my-subscription")
-                .subscribe();
-
-        Properties props = new Properties();
-        props.put("bootstrap.servers", lookupUrl.toString());
-
-        props.put("key.serializer", IntegerSerializer.class.getName());
-        props.put("value.serializer", StringSerializer.class.getName());
-
-        Producer<Integer, String> producer = new PulsarKafkaProducer<>(props);
-
-        CountDownLatch counter = new CountDownLatch(10);
-
-        for (int i = 0; i < 10; i++) {
-            producer.send(new ProducerRecord<Integer, String>(topic, i, "hello-" + i), (metadata, exception) -> {
-                assertEquals(metadata.topic(), topic);
-                assertNull(exception);
-
-                counter.countDown();
-            });
-        }
-
-        counter.await();
-
-        for (int i = 0; i < 10; i++) {
-            Message<byte[]> msg = pulsarConsumer.receive(1, TimeUnit.SECONDS);
-            assertEquals(new String(msg.getData()), "hello-" + i);
-            pulsarConsumer.acknowledge(msg);
-        }
-
-        producer.close();
-    }
-}
