+++ pulsar-common/pom.xml
@@ -68,8 +68,8 @@
     </dependency>
 
     <dependency>
-      <groupId>net.jpountz.lz4</groupId>
-      <artifactId>lz4</artifactId>
+      <groupId>org.lz4</groupId>
+      <artifactId>lz4-java</artifactId>
     </dependency>
 
     <dependency>
